requirements.txt
asgiref==3.8.1
cloudinary==1.44.1
Django==5.2.3
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
drf-yasg==1.21.10
pillow==11.2.1
psycopg2==2.9.10
whitenoise==6.9.0
